Pass http_proxy variables to sudo in run-root function
This is to ensure that sudo commands work from the beginnings on systems
that need to use a proxy. This will anyway be fixed by the script when
the custom sudoers file is installed, however, certain sudo commands
that need to access the internet, such as 'sudo apt update', are already
executed before the sudoers file is installed.

Once the sudoers file is installed, the hardcoded proxy variables don't
do any harm, they just become redundant because these variables are
already passed to the sudo environment through the sudoers file.