Add 'pipx' to packages and install Grip with pipx
The reason is that when using 'pip install grip' on Debian 12, there's
a "error: externally-managed-environment" error. This doesn't occur on
Ubuntu 22.04. See [1] for more information.

On Debian 12, this can be solved by adding the --break-system-packages
flag to 'pip install', however, the version of pip installed on Ubuntu
22.04 doesn't have this flag.

The installation method with pipx works on both systems.

[1] https://itsfoss.com/externally-managed-environment/