Remove requirement for Bash version >= 4.0
The script has been tested to work on Bash 4.0. However, it's not clear
what's the actual minimum required version as it's not easy to compile
and install older versions of Bash (e.g. Bash 3.2) from [1] on today's
platforms.

However, the Bash version will probably never be an issue as current
Linux distributions always include sufficiently new versions of Bash
(for comparison: Bash 4.0 is from 2009).

[1] https://github.com/bminor/bash